recommender/requirements.txt
pandas==2.2.1
pymongo==4.6.1
scikit-learn==1.4.1
streamlit==1.32.0
python-dotenv==1.1.0
